Big data bug fixes (#1074)
* Fixed the overflow error for the conic solver

I fixed an overflow error which was preventing the conic solver from solving large optimization problems. I am attempting to do the same thing for the MOSEK solver

* Reversing MOSEK changes to pull only conic solver fixes

I'm reverting my MOSEK solver fixes so that I can the only thing that gets pulled into the main branch is the conic solver overflow fix. This means this branch fixes an overflow error that was occuring on the shape check of the matrix. the datatypes of the variables were int32 which was causing the error. Changing these datatypes to int64 (the edits I made) allows larger optimization problems to be solved and no errors have been encountered.

+++ cvxpy/cvxcore/python/canonInterface.py
@@ -367,7 +367,7 @@ def get_problem_matrix(linOps,
     I = np.concatenate(I)
     J = np.concatenate(J)
     A = scipy.sparse.csc_matrix(
-        (V, (I, J)), shape=(constr_length*(var_length+1), param_size_plus_one))
+        (V, (I, J)), shape=(np.int64(constr_length)*np.int64(var_length+1), param_size_plus_one))
     return A
